Host image files on a separate server
This avoids the need to commit resized images to the git repo. It is
much more likely that the resized images will change than that the
sources will. We don't get back the space in git if we change the set of
images we're producing, so it's better for them to live elsewhere.