Fix fast linting configuration
Background:
We have fast linting with `.eslintrc.json` and slower, more thorough
linting with `.eslintrc.full.json` (which makes use of TypeScript type
information).

Issue and solution:
The fast linting was broken because `.eslintrc.json` contained settings
for a rule that is only defined in `.eslintrc.full.json`.  This commit
moves the setting for that rule into the other file.