Remove custom gulp plugin for running rollup
Now the gulp task returns a Promise from rollup's API.

Initially I implemented a transform stream for rollup in order to pipe
changes to browser-sync once it was added. However, the way that I ended
up implementing a browser-sync server was by simply watching the output
directory, which means that the server can identify which files were
updated without needing to stream the build task to the server. This
makes the custom gulp-rollup plugin unnecessary, since returning a
Promise from rollup's own API is much simpler.